make HB universe polymorphism "ready"
In that setting storing terms in a program is problematic, since
poly universes should be abstracted as clause parameters. We could
do that, but the simplest approach is to just store term names. It
is then the operation of "mentioning a name", building (global GR),
which will have to take care to instantiate a poly constant.
This stays todo, waiting for coq-elpi