wrap: wrapped subject is inferred [needs discussion]
When we build the wrapped we infer the subject.
If the subject has deps, it is inferred as the
sort projection of its canon instance satisfying
all deps.

For now I do a whd, but one should probably be less aggressive.

This needs discussion since subject are GIVEN not inferred, usually.